waterbox: 64-bit button mask, analog axes, and two per-frame cost fixes
ABI:
- FrameAdvance takes a 64-bit mask. 32 buttons does not cover a four-player
  console with peripherals; quickerNES alone needs 34.
- waterbox.config gains input.axes, and the adapter pushes each declared axis to
  the guest with the optional SetAxis(index, value) before the frame it belongs
  to. Analog values cannot ride in a button mask, and a paddle is not eight
  bools. A package that declares axes must export SetAxis.

Performance - rewind takes a savestate every frame, so anything per-state is
per-frame:
- SaveStateBinary/LoadStateBinary no longer bracket the host call with
  Deactivate/Activate. mb_host_save_state activates itself and restores what it
  found, so the bracket bought four full guest-address-space remaps per state.
- JoypadApi.Set(dict) ran the entire input pipeline once per button
  (LatchFromPhysical + Overrides inside the per-button setter) - quadratic in the
  definition size. It now records every button and latches once; the end state is
  identical, since the loop's last iteration is what survived.

New: joypad.setaxis, the analog counterpart of joypad.set. It pushes a value
through the override adapter for the current frame, which joypad.setanalog (a
sticky autohold) never does from a script - the reason the quickerNES witness had
to hand-build bk2 mnemonic strings to move a paddle.

With the miniBox host fix, 5000 frames of an NES core through the full frontend
went from 65.7s to 10.9s under rewind.

Witness: 17 ok, 0 failed.

+++ tests/synth/package-box/synth-run-box.c
@@ -41,7 +41,7 @@ static int32_t mem_write(uintptr_t ud, const uint8_t *d, uintptr_t n){mb_t*m=(mb
 static intptr_t mem_read(uintptr_t ud, uint8_t *d, uintptr_t n){mb_t*m=(mb_t*)ud;uintptr_t a=m->len-m->pos;if(n>a)n=a;memcpy(d,m->b+m->pos,n);m->pos+=n;return (intptr_t)n;}
 
 typedef int (*init_fn)(void);
-typedef void (*frame_fn)(uint32_t);
+typedef void (*frame_fn)(uint64_t);
 typedef uintptr_t (*getptr_fn)(void);
 
 static uintptr_t proc(mb_host *h, const char *n){mb_return r;wbx_get_proc_addr(h,n,&r);if(r.error_message[0]){fprintf(stderr,"proc %s: %s\n",n,r.error_message);exit(2);}return r.data;}
@@ -107,7 +107,7 @@ int main(int argc, char **argv) {
 			wbx_deactivate_host(h, &r); wbx_load_state(h, mem_read, (uintptr_t)&state, &r); wbx_activate_host(h, &r);
 			if (r.error_message[0]) { fprintf(stderr, "rerecord: %s\n", r.error_message); return 1; }
 		}
-		FrameAdvance(pad);
+		FrameAdvance((uint64_t)pad);
 		frames++;
 		sha1_update(&vh, (const void *)GetFramebuffer(), 128 * 120);
 		sha1_update(&ah, (const void *)GetAudio(), 735 * 2);
